feat: add proper focus navigation and animations
