+++ packages/gatsby-theme-carbon/src/styles/index.scss
@@ -11,7 +11,6 @@
 @import './internal/page';
 
 @import '../components/Homepage/homepage.scss';
-@import '../components/GlobalSearch/global-search.scss';
 @import '../components/PageTable/page-table.scss';
 @import '../components/DoDontExample/do-dont-example.scss';
 @import '../components/Caption/caption.scss';
